Not completely working yet

+++ MSP430/Booster_Pack_LoRa_SPI/driverlib/MSP430FR5xx_6xx/eusci_b_spi.c
@@ -118,7 +118,7 @@ void EUSCI_B_SPI_changeClockPhasePolarity (uint16_t baseAddress,
 }
 
 void EUSCI_B_SPI_transmitData ( uint16_t baseAddress,
-    uint8_t transmitData
+    uint16_t transmitData
     )
 {
     HWREG16(baseAddress + OFS_UCBxTXBUF) = transmitData;
